# Optimize inflightInstanceTracker