**What this PR does**:

Following up a [suggestion received here](https://github.com/grafana/dskit/pull/495#discussion_r1499277214), I've tried to replace the `map` used in `inflightInstanceTracker` with a slice and the performance impact is significative.

```
goos: darwin
goarch: amd64
pkg: github.com/grafana/dskit/ring
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                                                                    │  before.txt  │             after.txt              │
                                                                                    │    sec/op    │   sec/op     vs base               │
InflightInstanceTracker/few_replication_sets,_many_instances_per_replication_set-12   29.29µ ± 11%   31.08µ ± 7%        ~ (p=0.065 n=6)
InflightInstanceTracker/many_replication_sets,_few_instances_per_replication_set-12   49.89µ ±  8%   33.73µ ± 8%  -32.39% (p=0.002 n=6)
geomean                                                                               38.23µ         32.38µ       -15.31%

                                                                                    │  before.txt   │              after.txt              │
                                                                                    │     B/op      │     B/op      vs base               │
InflightInstanceTracker/few_replication_sets,_many_instances_per_replication_set-12    9.281Ki ± 0%   2.750Ki ± 0%  -70.37% (p=0.002 n=6)
InflightInstanceTracker/many_replication_sets,_few_instances_per_replication_set-12   13.422Ki ± 0%   5.016Ki ± 0%  -62.63% (p=0.002 n=6)
geomean                                                                                11.16Ki        3.714Ki       -66.72%

                                                                                    │ before.txt  │             after.txt             │
                                                                                    │  allocs/op  │ allocs/op   vs base               │
InflightInstanceTracker/few_replication_sets,_many_instances_per_replication_set-12   11.000 ± 0%   5.000 ± 0%  -54.55% (p=0.002 n=6)
InflightInstanceTracker/many_replication_sets,_few_instances_per_replication_set-12    202.0 ± 0%   102.0 ± 0%  -49.50% (p=0.002 n=6)
geomean                                                                                47.14        22.58       -52.09%
```

**Which issue(s) this PR fixes**:

N/A

**Checklist**
- [x] Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`